pyproject.toml
[project]
name = "gesture-recognition"
version = "0.1.0"
description = ""
authors = [
    {name = "Javi"}
]

requires-python = ">=3.10, <=3.10.14"
dependencies = [
    "adafruit-blinka (>=9.0.3,<10.0.0)",
    "adafruit-circuitpython-rgb-display (>=3.14.3,<4.0.0)",
    "board (>=1.0,<2.0)",
    "pillow (>=12.1.1,<13.0.0)",
    "st7735 (>=1.0.0,<2.0.0)",
    "gpiod (>=2.4.1,<3.0.0)",
    "gpiodevice (>=0.0.5,<0.0.6)",
    "mediapipe (==0.10.14)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false